refactor(query_api): extract git/GitHub/papertrail history surface to history.rs
Move commit_search, git_history_for_path/symbol, commits_touching_query,
git_blame_chunk, github_sync_*, github_issue_search, rationale_search,
github_refs_for_path, github_sync_status, and papertrail_for_* into history.rs.
Also restores run_oracle's doc comment, which the oracle_runs extraction left
behind in mod.rs (it had re-attached to status). mod.rs is now 753 lines (was
3900); query_api is 7 job-focused files. No behavior change.
