+++ crates/rag-rat-core/src/index/query_api/history.rs
@@ -1,0 +1,251 @@
+use super::*;
+
+impl IndexDatabase {
+    pub fn commit_search(&self, query: &str, limit: u32) -> anyhow::Result<Vec<CommitSearchHit>> {
+        git_history::commit_search(self.storage.connection(), query, limit)
+    }
+
+    pub fn git_history_for_path(
+        &self,
+        path: &str,
+        limit: u32,
+    ) -> anyhow::Result<Vec<PathHistoryItem>> {
+        git_history::history_for_path(self.storage.connection(), path, limit)
+    }
+
+    pub fn git_history_for_symbol(
+        &self,
+        symbol: &str,
+        language: Option<Language>,
+        limit: u32,
+    ) -> anyhow::Result<Vec<SymbolHistoryItem>> {
+        let symbols = self.symbols(symbol, language, limit)?;
+        let per_symbol_limit = limit.max(1);
+        let mut out = Vec::new();
+        for symbol_hit in symbols {
+            for commit in self.git_history_for_path(&symbol_hit.path, per_symbol_limit)? {
+                out.push(SymbolHistoryItem {
+                    symbol: symbol_hit.name.clone(),
+                    qualified_name: symbol_hit.qualified_name.clone(),
+                    path: symbol_hit.path.clone(),
+                    start_byte: symbol_hit.start_byte,
+                    end_byte: symbol_hit.end_byte,
+                    commit,
+                    evidence_kind: "historical",
+                });
+                if out.len() >= usize::try_from(limit).unwrap_or(usize::MAX) {
+                    return Ok(out);
+                }
+            }
+        }
+        Ok(out)
+    }
+
+    pub fn commits_touching_query(
+        &self,
+        query: &str,
+        limit: u32,
+    ) -> anyhow::Result<Vec<QueryCommitHit>> {
+        let current_hits = self.search(query, limit, true)?;
+        git_history::commits_touching_query(self.storage.connection(), query, limit, &current_hits)
+    }
+
+    pub fn git_blame_chunk(&self, chunk_id: i64) -> anyhow::Result<Option<ChunkBlameSummary>> {
+        let Some(chunk) = self.read_chunk(chunk_id)? else {
+            return Ok(None);
+        };
+        let source_text_hash = git_history::source_text_hash(&chunk.text);
+        if let Some(cached) =
+            git_history::cached_blame(self.storage.connection(), chunk_id, &source_text_hash)?
+        {
+            return Ok(Some(cached));
+        }
+        let Some(root) = self.storage.source_root() else {
+            return Ok(Some(ChunkBlameSummary {
+                chunk_id,
+                path: chunk.path,
+                start_line: chunk.start_line,
+                end_line: chunk.end_line,
+                source_text_hash,
+                line_count: 0,
+                dominant_commit: None,
+                dominant_commit_lines: 0,
+                newest_commit: None,
+                newest_commit_time_s: None,
+                oldest_commit: None,
+                oldest_commit_time_s: None,
+                commit_counts: BTreeMap::new(),
+                evidence_kind: "historical",
+            }));
+        };
+        let blame_lines =
+            git_history::blame_lines(root, &chunk.path, chunk.start_line, chunk.end_line);
+        let mut counts = BTreeMap::<String, i64>::new();
+        let mut newest = None::<(String, i64)>;
+        let mut oldest = None::<(String, i64)>;
+        for line in &blame_lines {
+            *counts.entry(line.commit.clone()).or_default() += 1;
+            if let Some(time) = line.author_time_s {
+                if newest.as_ref().is_none_or(|(_, newest_time)| time > *newest_time) {
+                    newest = Some((line.commit.clone(), time));
+                }
+                if oldest.as_ref().is_none_or(|(_, oldest_time)| time < *oldest_time) {
+                    oldest = Some((line.commit.clone(), time));
+                }
+            }
+        }
+        let dominant = counts
+            .iter()
+            .max_by_key(|(commit, count)| (*count, *commit))
+            .map(|(commit, count)| (commit.clone(), *count));
+        let summary = ChunkBlameSummary {
+            chunk_id,
+            path: chunk.path,
+            start_line: chunk.start_line,
+            end_line: chunk.end_line,
+            source_text_hash,
+            line_count: i64::try_from(blame_lines.len()).unwrap_or(i64::MAX),
+            dominant_commit: dominant.as_ref().map(|(commit, _)| commit.clone()),
+            dominant_commit_lines: dominant.map(|(_, count)| count).unwrap_or(0),
+            newest_commit: newest.as_ref().map(|(commit, _)| commit.clone()),
+            newest_commit_time_s: newest.as_ref().map(|(_, time)| *time),
+            oldest_commit: oldest.as_ref().map(|(commit, _)| commit.clone()),
+            oldest_commit_time_s: oldest.as_ref().map(|(_, time)| *time),
+            commit_counts: counts,
+            evidence_kind: "historical",
+        };
+        git_history::store_blame(self.storage.connection(), &summary)?;
+        Ok(Some(summary))
+    }
+
+    pub fn github_sync_from_refs(&self, offline: bool) -> anyhow::Result<GitHubSyncReport> {
+        self.github_sync_from_refs_with_progress(offline, |_| {})
+    }
+
+    pub fn github_sync_from_refs_with_progress(
+        &self,
+        offline: bool,
+        progress: impl FnMut(github::GitHubSyncProgress),
+    ) -> anyhow::Result<GitHubSyncReport> {
+        let Some(root) = self.storage.source_root() else {
+            anyhow::bail!("index has no source_root metadata; rebuild required");
+        };
+        if offline {
+            github::sync_from_refs::<github::GhCliGitHubClient>(
+                self.storage.connection(),
+                root,
+                None,
+                true,
+                &self.github,
+            )
+        } else {
+            let client = github::GhCliGitHubClient;
+            github::sync_from_refs_with_progress(
+                self.storage.connection(),
+                root,
+                Some(&client),
+                false,
+                &self.github,
+                progress,
+            )
+        }
+    }
+
+    pub fn github_sync_issue(
+        &self,
+        issue_ref: &str,
+        offline: bool,
+    ) -> anyhow::Result<GitHubSyncReport> {
+        if offline {
+            github::sync_issue::<github::GhCliGitHubClient>(
+                self.storage.connection(),
+                issue_ref,
+                None,
+                true,
+                &self.github,
+            )
+        } else {
+            let client = github::GhCliGitHubClient;
+            github::sync_issue(
+                self.storage.connection(),
+                issue_ref,
+                Some(&client),
+                false,
+                &self.github,
+            )
+        }
+    }
+
+    pub fn github_issue_search(
+        &self,
+        query: &str,
+        limit: u32,
+    ) -> anyhow::Result<Vec<GitHubEvidence>> {
+        github::issue_search(self.storage.connection(), query, limit)
+    }
+
+    pub fn rationale_search(&self, query: &str, limit: u32) -> anyhow::Result<Vec<GitHubEvidence>> {
+        github::rationale_search(self.storage.connection(), query, limit, &self.github)
+    }
+
+    pub fn github_refs_for_path(
+        &self,
+        path: &str,
+        limit: u32,
+    ) -> anyhow::Result<Vec<github::GitHubRef>> {
+        github::refs_for_path(self.storage.connection(), path, limit)
+    }
+
+    pub fn github_sync_status(&self) -> anyhow::Result<GitHubStatus> {
+        self.github_status()
+    }
+
+    pub fn papertrail_for_chunk(
+        &self,
+        chunk_id: i64,
+        limit: u32,
+    ) -> anyhow::Result<Option<Papertrail>> {
+        let Some(chunk) = self.read_chunk(chunk_id)? else {
+            return Ok(None);
+        };
+        Ok(Some(github::papertrail_for_chunk(
+            self.storage.connection(),
+            &chunk,
+            limit,
+            &self.github,
+        )?))
+    }
+
+    pub fn papertrail_for_symbol(
+        &self,
+        symbol: &str,
+        language: Option<Language>,
+        limit: u32,
+    ) -> anyhow::Result<Option<Papertrail>> {
+        let Some(symbol) = self.symbols(symbol, language, limit)?.into_iter().next() else {
+            return Ok(None);
+        };
+        Ok(Some(github::papertrail_for_symbol(
+            self.storage.connection(),
+            &symbol,
+            limit,
+            &self.github,
+        )?))
+    }
+
+    pub fn papertrail_for_selected_symbol(
+        &self,
+        symbol: &crate::query::symbol::SymbolHit,
+        limit: u32,
+    ) -> anyhow::Result<Papertrail> {
+        github::papertrail_for_symbol(self.storage.connection(), symbol, limit, &self.github)
+    }
+
+    pub fn papertrail_for_commit(
+        &self,
+        commit_hash: &str,
+        limit: u32,
+    ) -> anyhow::Result<Papertrail> {
+        github::papertrail_for_commit(self.storage.connection(), commit_hash, limit, &self.github)
+    }
+}
